+++ test/grid.jl
@@ -50,6 +50,14 @@
     @test all(iszero, (grid.m)::Array{Float32,3})
     @test all(iszero, (grid.x)::Array{Vec{3,Float32},3})
 
+    geometry = FEMesh(Tesserae.Quad9(), CartesianMesh(1, (0,2), (0,1)))
+    field = FEMesh(Tesserae.Quad4(), geometry)
+    field_grid = @inferred generate_grid(@NamedTuple{x::Vec{2,Float64}, p::Float64}, field)
+    @test field_grid.x === field
+    @test length(field_grid) == 6
+    geometry[1] = Vec(-1.0, -1.0)
+    @test field_grid.x[1] == geometry[1]
+
     # broadcast for SpArray
     grid = generate_grid(SpArray, @NamedTuple{x::Vec{2,Float64}, m::Float64, v::Vec{2,Float64}}, mesh)
     spinds = Tesserae.get_spinds(grid)